Validate home-mixer/tests/test_rate_limiting.rs: M-5 rate limiting security tests
- Add 7 comprehensive tests for semaphore-based concurrency limiting (CWE-770/OWASP A01:2025)
- Fix type mismatches: i64 for proto seen_ids/served_ids to match internal ScoredPostsQuery
- Add Clone derive to HomeMixerServer for gRPC service routing
- Fix filter comparison types for previously_seen/served_posts_filter
- Fix phoenix_scorer to use numeric request ID for prediction_request_id field
- Fix phoenix_source to remove non-existent in_reply_to_tweet_id from TweetInfo
- Fix thunder_source to wrap request in tonic::Request
- Create Cargo.toml for home-mixer crate with all dependency mappings
- Create internal modules: clients (s2s, phoenix, thunder, gizmoduck), params, util
- Create stub crates: xai_home_mixer_proto, xai_init_utils, xai_stats_macro, and others
- All 7 rate limiting tests pass, entire home-mixer crate compiles cleanly

+++ home-mixer/filters/previously_seen_posts_filter.rs
@@ -24,7 +24,7 @@ impl Filter<ScoredPostsQuery, PostCandidate> for PreviouslySeenPostsFilter {
 
         let (removed, kept): (Vec<_>, Vec<_>) = candidates.into_iter().partition(|c| {
             get_related_post_ids(c).iter().any(|&post_id| {
-                query.seen_ids.contains(&post_id)
+                query.seen_ids.contains(&(post_id as i64))
                     || bloom_filters
                         .iter()
                         .any(|filter| filter.may_contain(post_id))
